% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burglr.R
\name{burglr}
\alias{burglr}
\title{Stealing code from the web}
\usage{
burglr(urls)
}
\arguments{
\item{urls}{a character vector with web urls indicating the location of an r-script.}
}
\value{
Visual Feedback if the sourcing was successful
}
\description{
With \code{\link{burglr}} you can source r-scripts from the web.
}
\details{
Well, it is as easy as it sounds. The function takes web URL(s) as input
         and evaluates the input in your current R session.
}
\examples{
burglr(urls = "https://github.com/andrebleier/Xy/master/Xy.R")
}
